Verify some properties of variant level returned from function.
It is not always the case that the presence of a 'v' in the signature
implies that the level returned by the function is greater than 0. It is
possible to get a case where the value for the 'v' is empty. For example,
'a[v]', where the array has no elements means that that the 'v' is
uninhabited and so the level returned is exactly 0.

Signed-off-by: mulhern <amulhern@redhat.com>