Change way of getting devices
The devices are just strings that look like paths and have
nothing to do with real devices.

Require hypothesis to generate the strings, no longer
require pyudev to find the devices.

This is simpler and easier than getting actual devices.
Since the sim engine doesn't do anything with the devices, non-existant
devices are fine.

Signed-off-by: mulhern <amulhern@redhat.com>