Change stratis to explicitly use Python 3
Since we are Py3 only, we don't want to inadvertently fall back to Py2 if
that's the default Python interpreter on the system. This can result in
errors that are not obviously due to using the wrong Python. It's better
to just insist on the interpreter we need.

Alter README based on this.

resolves #55

Signed-off-by: Andy Grover <agrover@redhat.com>