Print subcommand-level help if no subsubcommand given
If a subcommand is given but no sub-subcommand, print the help for the
subcommand, not the top-level help. e.g. 'stratis pool' w/o giving a
pool subcommand prints the pool help.

The change is: we always set 'func'. If it's not given, we set it to print
help. (A lambda is used because we call func with the Namespace as
argument, whereas parser.print_help() doesn't want/need this.)

Also, we no longer need to set 'dest' for subparsers since it's no longer
checked.

However, we do want to pass a title parameter to add_subparsers. This seems
necessary to have 'subcommands' heading in help, rather than 'positional
parameters'.

Signed-off-by: Andy Grover <agrover@redhat.com>