Control raising of StratisCliError with --propagate flag
The intention of the --propagate flag is to make stratis appear tidier
to user, but to allow developers and more sophisticated users to get
the benefit of the stack trace.

Update all tests to set propagate flag so that StratisCliErrors can be
caught by the tests.

Add additional tests to verify that --propagate flag is respected.

Resolves: #66.

Signed-off-by: mulhern <amulhern@redhat.com>