Do not disable or ignore certain pylint errors
StratisdErrors needed to be ignored back when it was autogenerated from
information on the stratisd D-Bus API. Now that it's defined in a
straightforward way, it does not need to be ignored.

I don't know why no-self-use was needed originally, but it is not needed
now, so it might as well go away.

Signed-off-by: mulhern <amulhern@redhat.com>