Remove unused errors
Most of these are special marker errors from the time when the cli was
used as a way to talk about future funtionality of stratisd. Now that
stratisd is generally ahead of the CLI these can go away.

The StratisCliValueError wasn't useful and was replaced by more specific
errors a while ago.

Signed-off-by: mulhern <amulhern@redhat.com>