Do not disable bad-continuation warnings for pylint
Make a few edits so that the code will satisfy both pylint and yapf.

bad-continuation was disabled by me so that I could use my own formatting
style. Now that we are using yapf there is no point, so might as well
just conform to both. I think the code looks slightly better this way.

Signed-off-by: mulhern <amulhern@redhat.com>