Print a list of error messages instead of just the top one
If there is no printable string for any error message raise exception.
This will encourage users to report what should never happen.

Print the error messages out in a conversational way.
Start with the lowest level and move up to the highest level.
Interpolate "which in turn caused" in between error messages.

Include __context__ as well as __cause__ in chain of causes.
__context__ field is set if there is no "from" after the exception.
But that is just as likely to mean that the developer of the library
did not bother to use "from" than that from does not apply. In other words,
this seems like a backwards compatibility effort, more than anything, so
treat __context__ and __cause__ the same when reporting errors.

Put error reporting into a separate module. It is still quite limited,
but the plan is to make it more interesting, at least a module's worth.

Resolves: #83

Signed-off-by: mulhern <amulhern@redhat.com>