Pin dbus-python to 1.2.4
This allows running the lint test on Travis while Travis's version of
Ubuntu does not support libdbus v.1.8. This is the most recent available
version of dbus-python which is not the current, problematical one that
requires libdbus v.1.8. When Travis has upgraded sufficiently, this pin
can be dropped.

No longer allow lint task to be a failure.

Signed-off-by: mulhern <amulhern@redhat.com>