Group management of stratisd constants in _stratisd_constants
Modify value_to_name() so that its result takes an extra parameter,
depending on whether the format for an unknown value should be terse or
informative. The idea is that terse is better when printing out a table.

This slightly changes the display of some of the block dev states in the CLI:
* Not-in-use -> NotInUse
* In-use -> InUse

Signed-off-by: mulhern <amulhern@redhat.com>