Correct column alignment for unicode characters
When we calculate the column widths we need to use how many columns
the text will occupy, not the number of characters.  When we output
the text for each piece of the column we need to adjust the width
to compenstate for this as well.

Corrects: https://github.com/stratis-storage/stratis-cli/issues/205

Signed-off-by: Tony Asleson <tasleson@redhat.com>
Signed-off-by: mulhern <amulhern@redhat.com>