Do not supply an explanation for AttributeError
Let it propagate. With the new setup this is quite reasonable. If the
AttributeError is caused by a defect in our code, we will want to fix it
immediately. If it is caused by a defect in the code of one of our
dependencies, we will have to address the problem in some way or other.

Signed-off-by: mulhern <amulhern@redhat.com>