blackbox: make units_to_bytes return a string
The units_to_bytes function is always failing due to the input
variable "size" always containing a period (e.g.: "10.00"); the
call fails with "ValueError: invalid literal for int() with base 10".

However, all of the callers of units_to_bytes() only use the return
value as a string in a dictionary, therefore returning the value
as a concatenation of size + units allows the tests to pass.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>