Make calculation of objects lazy again
The calculation is still lazy in TopActions.list_pools, but was for some
reason made eager in:
35ebe4d2b7536c323e702012e7b44212dfd2e230 for blockdevs and subsequently
for filesystems.

The reasons for making the calculation lazy are still good. The search
function is itself lazy, unless require_unique is set; there's no reason
to rush eagerly on and make two separate lists when you only need one
to get your computation done.

Signed-off-by: mulhern <amulhern@redhat.com>