Add STRATIS_DBUS_TIMEOUT environment variable
The following functionalities have been implemented and tested:

- If the environment variable is not set, then the Stratis D-Bus call will be
made with the default timeout value of 120 seconds.
- If the environment variable is set to an acceptable value, as per the Stratis
man pages, then the Stratis D-Bus call will be made with the corresponding
timeout value.
- If the environment variable is set to a non-acceptable value, then the
Stratis D-Bus call will not be made and the user will be provided feedback
about why the timeout value is not acceptable.

To support the above:

- A new subdirectory, unittest, has been added to the
tests/whitebox directory. This subdirectory contains the timeout tests.
- A new Makefile target, unittest-tests, has been added to the Makefile.
This target runs the tests in the unittest directory.
- A new Travis task, test-travis, has been added. This task runs the
tests in the unittest directory.
- The man pages have been updated. The timeout value specifications
are detailed in docs/stratis.txt.