Make blackbox directory not contain a package
This interacts poorly with the mechanism for running the tests, causing
the tests to not find the testlib module. Rather than change the tests so
that they are run in a different way, adapt the CI so the tests can be
run as they were run previously and pylint can check all the code.

Signed-off-by: mulhern <amulhern@redhat.com>