Add a test for KeyboardInterrupt exception
It uses monkey-patching, so put it in a separate directory and run it
separately.

Remove the related but very different test in stratis_cli_cert.py. It could
fail when it should succeed and it could also succeed when it should fail.
It could fail, if the machine was very slow and the interrupt was sent
to the process while the Python interpreter was coming up and before it had
started executing stratis code. It has a sleep call to make that less
likely, but a hard-coded time value could always be wrongish. It could
succeed when it should fail if there was some other error in the call,
before the interrupt signal was caught, which stratis had handled properly.
In that case, the return code would still be non-zero, and there would
be no traceback, but nothing would have been proved about how stratis
handles keyboard exceptions.

Signed-off-by: mulhern <amulhern@redhat.com>