Import generated classes in the methods in which they are used
In this way, the processing of the _data.py module is deferred until
stratis is executing.

The advantage of this is that all code generation is deferred until
the code is needed. In this way, any exception raised during code
generation is raised after all errors in the _errors.py module have been
defined.

Signed-off-by: mulhern <amulhern@redhat.com>