Make new errors a subtype of RuntimeError
Reword description of RuntimeError so that this organization makes more
sense.

Signed-off-by: mulhern <amulhern@redhat.com>

Change header comments for No- and Partial- change errors

Try to be more formal and avoid past tense descriptions of stratisd's
behavior because it may be that the exception is raised without any request
of stratisd to perform the action.

Signed-off-by: mulhern <amulhern@redhat.com>

Use StratisCli- naming convention for new exceptions

Signed-off-by: mulhern <amulhern@redhat.com>

Make StratisCliNoChangeError a subtype of StratisCliPartialChangeError

Have StratisCliNoChangeError's initializer invoke
StratisCliPartialChangeError's. This way, instances of both classes have
the identical attributes with the identical meanings, and the Liskov
Substitution Principle is not violated. In Java, the same effect would
be achieved by overloading the constructor of the
StratisCliPartialChangeError, but here it seems better to use subtyping.

The motivation is that, in all cases where StratisCliNoChangeError is
raised, there is no change, but this also can be true in cases where
StratisCliPartialChangeError is raised.

Do not match on StratisCliNoChangeError in _error_reporting.py, but just
on StratisCliPartialChangeError. Distinguish between no change and
partial change by means of the newly introduced partial() method.

Change StratisCliPartialChangeError's __str__ method for singular and
plural numbers of resources. This is kind of trivial, but users will
complain if we do not.

Change "command" to "action" in StratisCliPartialChangeError message
string, probably the command is the CLI command, and that is not what is
being reported here. Also, avoid past tense and "completed", the current
state could have been arrived at in many ways.

Document the invariant that unchanged_resources must not be empty; if it
were there would be no point in raising the exception.

Tidy up param specs in StratisCliPartialChangeError's docstring.

Signed-off-by: mulhern <amulhern@redhat.com>