Put no cover pragmas on most engine error branches
We should not be able to reliably elicit an engine error anywhere in the
tests, except when attempting to destroy a pool w/ filesystems.

Currently, it is possible to elicit an engine error in other circumstances,
because some parts of the D-Bus API are not idempotent. That will change
when the work on idempotency is merged, and all branches that raise a
StratisCliEngineError will not be covered, except for pool destroy.

Boost test coverage requirement to 98%.

Signed-off-by: mulhern <amulhern@redhat.com>