Require dbus-python-client-gen 0.7
It was released in May 2018 upstream and in Fedora, and so we can
expect that stratis-cli has been being run using it for over a year.
If we require it, we can get exceptions w/ better information, which we
can then use in _error_reporting.py to give better information to users.

The only other changes in the version were refactorings to reduce the
nesting of closures in the implementation and minor formatting and pylint
tidies.

Signed-off-by: mulhern <amulhern@redhat.com>