Remove partial() method from StratisCliPartialChangeError
Its only use is in error reporting, where it allows to decide between
two error strings. But the error string actually belonging to
StratisCliPartialChangeError does an excellent job of precisely indicating
what would have happened, so the extra effort in error reporting is
redundant.

Note that the elimination of partial() also brings test_coverage in
_errors.py up to 100%.

Signed-off-by: mulhern <amulhern@redhat.com>