Replace erroneous StratisCliNoChangeError
If DestroyPool is invoked, that must mean that the object path for the
pool was obtained and therefore the pool exists. If DestroyPool reports
no change, then either the pool was removed in between D-Bus calls, or
there is a bug in stratisd. Therefore it is appropriate to raise a
StratisCliIncoherenceError rather than a StratisCliNoChangeError here.