Call "udevadm settle" once setup for each individual test is completed
Everybody knows that calling udevadm settle is never ideal, because this
means that the process must wait until all events in the udev queue have
been processed, even ones completely unrelated to the action that is
desired to be taken. However, there is no obvious better solution at this
time.

Also, use clean_up() in StratisCertify.setUp() in stratisd_cert.py
as well as in stratis_cli_cert.py.

Signed-off-by: mulhern <amulhern@redhat.com>