Calculate physical size triple to display for pool list
Generalize fetch_property so that the result can have any type. This
requires returning None if there was a failure, and then processing None
after each indvidual call to fetch_property, which is somewhat verbose
but may be the correct thing to do, in order to accomodate the required
generality. Use a constant for the designated string to use to represent
failure, "FAILURE".

Change "Total Physical Size" column header to "Total Physical", since it
now displays a three part value, size/used/free.

Signed-off-by: mulhern <amulhern@redhat.com>