Refuse to try to create a pool if there is a name conflict
We have decided that this will probably be a simpler experience for the
CLI user than allowing them to see the set of interesting and more
nuanced responses that the D-Bus API provides in different situations.

Corrected a test header comment at the same time as the test was updated
for these changes.

Signed-off-by: mulhern <amulhern@redhat.com>