Also run format tasks on bin/stratis
Previously, the fmt task was skipping the bin directory.

isort doesn't do anything to the imports in bin/stratis because they
are within some code that causes isort to ignore them. But run it anyway,
because this will change eventually, either because of a change to isort
or because of a change to our code.

Signed-off-by: mulhern <amulhern@redhat.com>