Expand test coverage
The __str__ method for Stratis Action Error is invoked, which allows
our coverage to be expanded.

Additionally, a justification for not covering
_interface_name_to_common_name is added, as there is currently no action
to specify a particular block device, once it has been added to a pool.
Once this type of action is possible, testing may be expanded to cover
this case.

Signed-off-by: Leah Leshchinsky <lleshchi@redhat.com>