Use a main method in blackbox test scripts
Make the module-level global DISKS a class attribute instead; its behavior
is better controlled that way.

Rename OTHER_ARGS to unittest_args. Any args that are not "--disk"
args will be picked up and passed as a list to the unittest main method.

Do not make the "--disk" "option" required. The required flag has the
effect of returning an error during parsing when the "--disk" option is
not used at all, but since the minimum number of disks required to actually
run all the tests is 3, requiring the option to be used 3 times,
making it required doesn't force the user to call the script correctly.

Document the requirement of at least 3 disks in the help text.

Signed-off-by: mulhern <amulhern@redhat.com>