Move the _get_timeout function out of _data.py
It's nice to be able to test it w/out making all those classes.

Freshen up the equivalent method in the blackbox tests and explain why
it's a copy not an import. Use RuntimeError instead of ValueError;
I would like it if only the interpreter raised ValueError, and I want to
make it so here, at least.

Signed-off-by: mulhern <amulhern@redhat.com>