Have fetch_property raise an exception if an error was encountered
Create a new exception for this purpose and handle it in
_error_reporting.py.

Create a new method, get_property, in _formatting which returns a default
value if fetch_property raises an exception. Use this wherever
fetch_property was previously used.

Fix all call sites and fix an unlikely bug in generating
strings for boolean properties. Previously it was assumed that the only
possibilities were True and False. Now allow for unknown and indicate that
with a "?" prefix. Note that in the particular case where this is handled,
an unknown value is enormously unlikely, but we may in future handle other
values where this is less unlikely.

Move fetch_property into _utils; it's not just about formatting anymore.

Add a test to cover newly introduced branch.

Update man pages to include mention of the "?".

Revise comments and error message for StratisCliPropertyNotFoundError.

Signed-off-by: mulhern <amulhern@redhat.com>