Remove interface name from fetch_properties
Also remove it from related exception types.

This information is useful, but in the general case it is not really
guarateed to be available when this method is invoked. Missing information
should be handled by exception chaining. But most likely very few people
will ever see either of these errors.

Signed-off-by: mulhern <amulhern@redhat.com>