Distinguish between connection to D-Bus and stratisd running failure
The "org.freedesktop.DBus.Error.NameHasNoOwner" error can arise for two
reasons: either the stratisd process is not running or there was a
failure connecting to the D-Bus. It is useful to report the likely cause
of the error, and the two can be distinguished by using psutil to check
whether a stratisd process is running.

Signed-off-by: Leah Leshchinsky <lleshchi@redhat.com>