Use LockedPools property in unlock actions
Since the new FetchProperties interface supplies LockedPools, and we
therefore expect that eventually LockedPoolUuids will be deprecated, we
might as well move up now.

The returned value is a dict rather than a list, so change the
equality test. Change the identifier, too, it becomes a little deceiving.
Do not alter the for-loop, Python overloading means that it is interpreted
as getting the keys from the dict rather than the items from the list.

Signed-off-by: mulhern <amulhern@redhat.com>