Change file descriptor passed to stratisd in interactive mode
This commit changes the interactive mode to provide stratisd with the
read end of a pipe. This has a few benefits over providing stdin
directly to stratisd. One benefit is that stratis-cli can now terminate
a stratisd read by closing the write end of the pipe which the CLI has
full ownership over. This also works around the SELinux rule that
prevents us from passing a file descriptor pointing to a TTY to a
daemon.