Drop dependency on psutil
Prior to the last release, stratisd would retry getting a connection to
the D-Bus and so there might be an appreciable or even infinite time
when stratisd could be up, but it had not made a D-Bus connection.

Now, if it doesn't make a D-Bus connection, it fails immediately.

Since the condition the error checks for has a much reduced probability,
we can drop the dependency on psutil altogether.

We don't drop psutil dependencies in GitHub workflows, psutil is still
used in tests, which means that is required for tests and the linter
has to know about it also.

Signed-off-by: mulhern <amulhern@redhat.com>