Add secret command-line option to print all help text
Suppress the help text for this option.

In order to find all subcommands, the code needs to introspect on the
parser, which is most easily achieved within stratis-cli source.

Canonicalize the order of the help text output alphabetically by
subparser name.

Add a test for the --print-all-help option and remove the old test because
it duplicates the --print-all-help option action.

Suppress the output from the help text unit test. It's fairly verbose and
after this PR the output will be available from the stratis-cli
command-line, so there will be almost no utility in having it stored in
every test run also.

Signed-off-by: mulhern <amulhern@redhat.com>