Put UUID/name parser chunk in _shared.py
Consistently put name first and UUID second.

We use this construction in many places, it seems a shame to repeat it.

This change means that everywhere the UUID/name alternative is displayed
in the help text, it will not have a help message specific to the
operation, but will instead have a generic help message as in this
example:

usage: stratis pool list [-h] [--stopped] [--uuid UUID | --name NAME]

options:
  -h, --help   show this help message and exit
  --stopped    Display information about stopped pools only.
  --uuid UUID  UUID
  --name NAME  name

whereas previously it would have looked like this, with specific help
messages for the uuid and name options.

usage: stratis pool list [-h] [--stopped] [--uuid UUID | --name NAME]

options:
  -h, --help   show this help message and exit
  --stopped    Display information about stopped pools only.
  --name NAME  name of pool to list
  --uuid UUID  UUID of pool to list

Signed-off-by: mulhern <amulhern@redhat.com>