Fix a bug where a stopped pool could not be started
This would happen if there was no token slot specified, an unlock method
of "clevis" or "keyring" was specified, and the stopped pool was
designated by its UUID rather than its name.

The command-lines:
> stratis pool start --uuid=<UUID> --unlock-method=keyring
> stratis pool start --uuid=<UUID> --unlock-method=clevis

If the pool is a V2 pool, then the only change is that the user receives
a different error, that the pool doesn't exist instead of that the
operation is invalid for that pool.

If the pool is a V1 pool, though, the bug would prevent it being
started, since the pool would also appear to not exist in this case.

Signed-off-by: mulhern <amulhern@redhat.com>