Allow unrecognized pool features
Add environment variable to enforce strictness if necessary.

Do not add a command-line option to override the environment variable,
as this will just introduce clutter.

Returning a None from the _missing_() method causes a ValueError to be
raised if the PoolFeature constructor argument is non-matching.

Enforce strictness in the CI by setting STRATIS_STRICT_POOL_FEATURES=1
in the GitHub CI.

Signed-off-by: mulhern <amulhern@redhat.com>