Use checkout/@v4 to check out stratisd repo
We can be consistent with how we check out the stratis-cli repo.

Also, the structure makes it easier to fill in a different ref, e.g.,
the ref for a pull request on the stratisd repo.

The only trick, then, is to figure out what the pull request should be.

Signed-off-by: mulhern <amulhern@redhat.com>