Use isoparse specifically to parse Created property
The value is generated by DateTime::to_rfc3339_opts; that function's
documentation states explicitly that the format is
   "RFC 3339 and ISO 8601 date and time string with subseconds formatted "
   "as per SecondsFormat"

The isoparse method documentation states explicitly that it parses
   "an ISO-8601 datetime string into a datetime.datetime."

This should always be correct.

It should also run about six times as fast, although stratis-cli is
unlikely to get any observable benefit from this.

>>> timeit.Timer("parser.isoparse('2025-05-01T13:13:26.863109')",
                 "from dateutil import parser").timeit()
8.737638719998358
>>> timeit.Timer("parser.parse('2025-05-01T13:13:26.863109')",
                 "from dateutil import parser").timeit()
54.581701957002224

Signed-off-by: mulhern <amulhern@redhat.com>