Half D-Bus timeout
We had it as high as two minutes to give a chance of returning on fairly
long-running task, like creating encrypted pools, since each device had
to be separately encrypted. We do not do that anymore, as the whole pool
is encrypted these days, so the create method returns faster.

We are about to introduce really long running commands, where 120
minutes will not be enough in almost all cases.

So shortening the D-Bus timeout seems like a reasonable thing to do.

60 seconds is a plenty long time to wait for any error messsages that
stratisd might return.

Signed-off-by: mulhern <amulhern@redhat.com>